Add confirmation box when deleting image report.
Fixes bug 1464 (https://bugs.linaro.org/show_bug.cgi?id=1464).

(cherry picked from commit 15f27460c6a72ceb355ce678762a66f15b414a5d)

+++ dashboard_app/views/image_reports/tables.py
@@ -51,7 +51,7 @@ class UserImageReportTable(LavaTable):
     view.orderable = False
 
     remove = tables.TemplateColumn('''
-    <a href="{{ record.get_absolute_url }}/+delete">remove</a>
+    <a href="{{ record.get_absolute_url }}/+delete" onclick="return confirm('Are you sure you want to delete this Image Report?');">remove</a>
     ''')
     remove.orderable = False
